requirements.txt
aiofiles==24.1.0


azure-storage-blob==12.23.1
azure-identity==1.19.0
azure-ai-formrecognizer==3.3.3
azure-search-documents==11.5.2
azure-core==1.34.0
azure-cosmos==4.9.0
azure-cognitiveservices-speech==1.34.1
azure-ai-contentsafety==1.0.0

nest_asyncio==1.6.0
psutil==7.0.0
PyPDF2==3.0.1
openai==1.75.0
crewai==0.121.0
crewai-tools==0.45.0
python-dotenv==1.0.1

imagehash==4.3.1
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
open-clip-torch==2.32.0
scikit-learn==1.5.2

fastapi==0.115.9
uvicorn[standard]==0.34.0
jinja2==3.1.6


passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
pillow==11.2.1

python-multipart==0.0.20

email-validator==2.2.0
itsdangerous==2.2.0    

requests>=2.31.0
pydub>=0.25.1
ffmpeg-python>=0.2.0